Refactor: renaming environments
