+++ experiments/training/td3.py
@@ -15,7 +15,7 @@ from tf_agents.metrics import tf_metrics
 
 from optfuncs import tensorflow_functions as tff
 
-from src.environments import tf_function_environment as tf_fun_env
+from src.environments import tf_function_env as tf_fun_env
 from src.metrics import tf_custom_metrics
 from src.typing.types import LayerParam
